Add example extensions to catalog and my collections
This isn't a perfect example of extending the the Sufia/Blacklist/Hydra
concepts, but does offer something to hack on. There are two new
controllers with index views that retrieve the collections like the
standard actions. That is, there is a view for everything and a view for
my collections. These just call up to super in the controller and
provide a vanilla view. Renaming or removing these views will use the
parent view directly from Sufia.

 * ItemsController extends CatalogController
 * My::ItemsController extends My::CollectionsController

The Gemfile uses a relative path to Sufia, but the example does not
depend on any modifications therein. The Sufia source can be modified at
runtime and most changes will be reflected without restarting.

The pry, pry-rails, and pry-nav gems have been added to aid in
inspecting objects or flow. For example, adding binding.pry after the
super call in my/collections_controller.rb#index in the Sufia source
would give a way to inspect the @response or @document_list. There are
more sophisticated debugger solutions available, but this can provide
good insights with minimal tooling. Extending the above example, having
at least one collection and loading /my_items will trigger a breakpoint
in the rails server process and the following code will show the methods
and properties available on the first collection, revealing things like
the to_model method and interactively exploring:

  ls @document_list.first
  cd @document_list.first.to_model
  title